requirements.txt
beautifulsoup4==4.13.4
google-generativeai==0.8.5
grpcio-status==1.71.0
pip==25.1.1
python-dotenv==1.1.0
